Changing `djangae.db.backends.appengine.query.Query.columns` from a `list` to a `set` to avoid duplicates.
This fixes the issue where doing `.distinct(x).values_list(x)` would cause an error due to the same column being projected twice.

Updating tests accordingly.

It seems slightly odd that we don't need `columns` to be ordered, but we don't seem to.  It doesn't change the number of expected failures (which we automatically mark) in the Django tests.