Refactoring unique marker logic.
This changes the way that we do the checking/creation/deletion of the UniqueMarker objects in the commands so that we start a transaction, Put() the entity, and THEN set the markers in an independent, nested transaction.  So if the marker setting raises an error then the outer transaction that does the Put() is rolled back.