Add djangae.environment, deprecate utils functions
This change adds djangae.environment which contains utility functions for
detecting information about the running environment. As part of this change
some of the functions that have lived in djangae.utils have been moved and
renamed to better names. The djangae.utils functions will still work for now
but will be removed before 1.0 and now emit deprecation warnings when used.