Change the flow of try/exept in `ContextDecorator`.
Previously, if there was an error in `_do_enter` then `_do_exit` would still be called.  And given that in the `AtomicDecorator`, `_do_exit` relies on things that `_do_enter` does, this means that another error would be raised in `_do_exit`, and the original error from `_do_enter` would never be seen, making debugging extremely difficult.