Don't depend on apps which don't have migrations.
Recently we added a clever hack to patch our simulated contenttypes manager even when
running migrations. This was necessary to get Wagtail working on GAE.

Our hack works by creating a migration which is a dependency of everything, however when calculating
the app list we were taking into account apps which didn't even have migrations. This commit
fixes that by ignoring apps which don't have a migrations module, or have on without files.