Updated the runserver management command to use argparse instead of optparse.
Django 1.10 has removed all use of optparse, so Djangae's runserver command
died with an AttributeError exception. This change is compatible with
Django 1.8 and 1.9.

Also I cleaned up the imports to follow the recommended style, and changed
the _get_sandbox_options helper to a static method because it doesn't use
the class at all.