Switch to using google.appengine.api.memcache to use compare-and-set
This commit makes the steal_after_ms argument work the same in memcache as the datastore. We perform
a compare-and-set check to deal with multiple threads trying to gain the lock.

Although it would be nice to use django.core.cache, this doesn't expose CAS operations (you can get to them
but through a private API so it's safer to just use memcache directly).

Using memcache also is safer as it won't use the Django KEY_PREFIX (which is often application-version specific) and
so will work when using multiple versions of the app simulaneously.