Changes the session_csrf security check to check the MIDDLEWARE setting.
Django 1.10 introduced a new MIDDLEWARE setting, and session_csrf works
fine with it. This check will work with pre-1.10 projects which don't
have the MIDDLEWARE setting defined at all, with 1.10 projects which
use MIDDLEWARE in place of MIDDLEWARE_CLASSES, and with 1.10 projects
which use MIDDLEWARE_CLASSES but not MIDDLEWARE (in which case the
default for MIDDLEWARE is `None`).

This fixes #857.