Change projections by key to use AsyncMultiQuery
There is a slight change in behaviour here. Before we used to run multiple, multi-queries when
a projection-by-key happened. Now we'll only run a MultiQuery if we can do it in a single one,
otherwise we fall back to datastore.Get. This means if you are getting a lot of results
your query won't be a projection, just a Get.