Moved import of Users API inside the view decorator.
This avoids an ImportError exception at startup. The sandbox pulls in
djangae.environment, and that was importing google.appengine.api before
the sandbox had a chance to put the SDK on the path (for those cases
where a project is using the location of dev_appserver.py to have the
SDK added to sys.path).