A test that importing djangae.urls doesn't import optional stuff.
This test requires the mock library because I got bored trying to
nicely patch sys.modules without it. And if the test didn't patch
sys.modules nicely, then a lot of unrelated tests complained about
import errors (because the test explicitly messes with imports to
make sure that the optional modules are _not_ imported).