Update `MapReduceEntityTests.test_filters` to test the filtering.
This is a slightly horrible change, but as `map_reduce_entities` doesn't allow us to pass any params to the mapper function, we can't use the `count` function.  Updating `map_reduce_entities` to allow passing params to the mapper function involves way too much investigation into the inner workings of the 'mapreduce' and 'pipeline' libraries, and we can't `sleuth.watch` or `sleuth.patch` anything because sleuth's `wrapped` objects cannot be pickled.