Allow computed fields to take the name of a method instead of a function object as the "computer" function.
The main benefit of this is that it allows the field to be pickled, even when using a method on the model class as the computer function.  Unbound methods on the class can't be pickled, so if you pass in a method directly into the field, you then can't pickle the field.