Fixed an `ImportError` at startup when the SDK isn't on the Python path.
The current logic allows one to have the SDK tools (dev_appserver.py, etc.)
on one's $PATH, and then Djangae works out where the SDK is installed and
adds it to sys.path, after which it is OK to import things from the SDK
modules.

But that got broke because `./manage.py runserver` indirectly loads this
module, which was pulling in things from the SDK. If you don't have the
SDK already already on sys.path, this would result in ImportError.