Changed the logic for generating keys when sharding.
We had a bug where keys that were very high or very low in the range of
valid keys would be ignored when iterating over entities (so the data
migration was not being applied for all entities).

The fix is to make sure that we always start one shard with the very
lowest key ID, and that the last shard has an unbounded upper range
for keys.

Needs more tests.

Fixes #1060.