Fixed test for number of shards in migrations.
The test broke because with 20 entities and `shard_query(q, 20)` the
old code returned 20 shards, while the new code returns 12. The old
code made effort to generate additional shards when the initial
key query didn't return enough keys. But new code doesn't, and the
scattered query for 20 * 32 keys when there are only 20 entities
actually returns 10 keys, to which we append the first and last keys
to bring the total to 12.