Add a `break` statement into `current_transaction`
This doesn't actually affect the behaviour because no other transaction will exist, we were just pointlessly continuing to loop.

Also removed `active_transaction = None` because all cases where we set this variable are now followed by a `break`, so if we get to this line then it will always be `None` anyway.