break after removing add_branch
After determining the add_branch should be removed while resolving detect_conflicting_key_filter, we should stop trying to remove the add_branch. It doesn't matter how many times we confirm it should be removed, we should just remove it once to avoid the ValueError that occurs from trying to remove it again.